sync

+++ frontend/src/store/task/task.action.js
@@ -8,7 +8,7 @@ export async function saveTask(task) {
         const { board } = store.getState().boardModule
         const savedTask = await taskService.save(task)
         const group = board.groups.find(group => group._id === task.groupId)
-        console.log('group:', group);
+        console.log('group:', group)
         if (task._id) {
             group.tasks = group.tasks.map(currTask => currTask._id !== savedTask._id ? currTask : savedTask)
         }
@@ -24,6 +24,7 @@ export async function saveTask(task) {
 }
 
 export async function removeTask(groupId, taskId) {
+    if (!groupId || !taskId) return
     try {
         const { board } = store.getState().boardModule
         const removedTaskId = await taskService.remove(taskId)
